// CLIENT/src/interfaces/validateMessage.ts
export enum EValidateMessage {
    required = "Please fulfill this field",
    date = "Please choose a date",
    phone = "Phone is invalid",
    email = "Email is invalid",
    password = "Password is incorrect",
    gender = "Please choose your gender",
    whiteSpace = "should not have whitespace character"
}